requirements.txt
numpy
pandas
fastapi
uvicorn
sentence-transformers==4.0.2
faiss-cpu==1.10.0
pydantic
openpyxl
